(WIP) Unifies lookup via tupling
This introduces a declaration(_)inContext(_) tuple and uses that to
unify lookups for declaration (aka invocable) and declarning context.
Methods already know their defining context, and that is always the
same as the declaraing context found by the lookup as far as I can
tell (see assert in the tuple).

So, plan now is to rip all that out, make every invocable know
its defining context, and just that when and where necessary.