Starts building "static" typechecking dialect.
I made some good progress, but had to abaondon this because there is apparently
no representation of an identifier in the "common-ast" — applied occurences of
identifiers are represented as unresolved "implicitRequests".  Thus, there is
no simple way for a type-cyhecking dialect to, for example, to find out
the declared type of a parameter reference.

The "common-ast" is in fact just a representation of the raw syntax, without
any of the analysis that a dialect-writer would need.  Using this ast, the
dialect write is going to have to reproduce large parts of the compiler.

It might be possible to have a method `resolve` on each unresolved request
that would invoke the compiler logic to resolve it, and return a oneself
request, or indentifier reference.

Going ofrward,it makes more sense to abandon this fork and move the type-
checker to minigrace.