Adds basic annotation and brand checking
Looks at the first annotation on object constructors
which can currently only come from a class via
a nastily hacked Kernan.  If there is something
there we eval it, if the resulting object has
a method called retainedAnnotation then we
add the object in as key to the objectContext
we are constructing.     A builtin method
can check for identity matches of objects
and brands (primitiveBrandMatch) - probably
that should be on builting brands, but
actually I don't think it breaks security
because you have to have the brand (capability)
to be able to call it in the first place.